(PC-17814) identityCheck: separate useSubscriptionNavigation helpers into smaller helpers
