+++ src/features/identityCheck/pages/helpers/useSubscriptionNavigation.ts
@@ -4,31 +4,14 @@ import { useQueryClient } from 'react-query'
 
 import { usePatchProfile } from 'features/identityCheck/api/usePatchProfile'
 import { useSubscriptionContext } from 'features/identityCheck/context/SubscriptionContextProvider'
+import { getNextScreenOrStep } from 'features/identityCheck/pages/helpers/getNextScreenOrStep'
+import { isSubscriptionRoute } from 'features/identityCheck/pages/helpers/isSubscriptionRoute'
+import { useCurrentSubscriptionStep } from 'features/identityCheck/pages/helpers/useCurrentSubscriptionStep'
 import { useSubscriptionSteps } from 'features/identityCheck/pages/helpers/useSubscriptionSteps'
-import { SubscriptionScreen, IdentityCheckStep, StepConfig } from 'features/identityCheck/types'
-import { subscriptionRoutes } from 'features/navigation/RootNavigator/subscriptionRoutes'
-import {
-  SubscriptionRootStackParamList,
-  UseNavigationType,
-} from 'features/navigation/RootNavigator/types'
+import { IdentityCheckStep, NextScreenOrStep } from 'features/identityCheck/types'
+import { UseNavigationType } from 'features/navigation/RootNavigator/types'
 import { eventMonitoring } from 'libs/monitoring'
 import { QueryKeys } from 'libs/queryKeys'
-
-type NextScreenOrStep = { screen: SubscriptionScreen } | { step: IdentityCheckStep } | null
-
-const isSubscriptionRoute = (name: string): name is SubscriptionScreen =>
-  subscriptionRoutes.map((route) => route.name).includes(name as SubscriptionScreen)
-
-const getCurrentStep = (steps: StepConfig[], currentRoute: keyof SubscriptionRootStackParamList) =>
-  steps.find((step) => step.screens.includes(currentRoute)) || null
-
-const useCurrentSubscriptionStep = (): IdentityCheckStep | null => {
-  const { name } = useRoute()
-  const steps = useSubscriptionSteps()
-  const currentRoute = isSubscriptionRoute(name) ? name : null
-  const currentStep = currentRoute ? getCurrentStep(steps, currentRoute) : null
-  return currentStep ? currentStep.name : null
-}
 
 const useNextScreenOrStep = (): NextScreenOrStep => {
   const { name } = useRoute()
@@ -78,26 +61,4 @@ export const useSubscriptionNavigation = (): {
     },
     isSavingCheckpoint,
   }
-}
-
-export const getNextScreenOrStep = (
-  steps: StepConfig[],
-  currentRoute: SubscriptionScreen | null
-): NextScreenOrStep => {
-  if (!currentRoute) return null
-  const currentStep = getCurrentStep(steps, currentRoute)
-  if (!currentStep) return null
-
-  // Step is not completed
-  const currentScreenIndex = currentStep.screens.indexOf(currentRoute)
-  const isLastScreen = currentScreenIndex === currentStep.screens.length - 1
-  if (!isLastScreen) return { screen: currentStep.screens[currentScreenIndex + 1] }
-
-  // Step is completed => find next step
-  const currentStepIndex = steps.indexOf(currentStep)
-  const isLastStep = currentStepIndex === steps.length - 1
-  if (!isLastStep) return { step: steps[currentStepIndex + 1].name }
-
-  // Step is complete and is last step
-  return { step: IdentityCheckStep.END }
 }
