--- src/features/identityCheck/pages/phoneValidation/errors/index.tsx
@@ -1,2 +1,0 @@
-export * from './PhoneValidationTooManyAttempts'
-export * from './PhoneValidationTooManySMSSent'
